created assigmentlist component
